Modified the finance_calculator.py file to comply with requirements

+++ python_introduction/finance_calculator.py
@@ -3,7 +3,7 @@ monthly_income = input("Enter your monthly income: ")
 monthly_expenses = input("Enter your total monthly expenses: ")
 
 #Calculating Monthly Savings
-monthly_savings = int(monthly_income) - int(monthly_expenses)
+monthly_savings = float(monthly_income) - float(monthly_expenses)
 
 #Projecting Annual Savings
 simple_annual_interest_rate = 0.05
